SSCS-2638 Minor Refactor
Move Integration tests to the Integration module
